requirements.txt
google-api-python-client
google-auth
pandas
openai
python-telegram-bot==20.0
